Convertet from Typscript to Javascript, WiFi Panel Heater driver and device
